added locality

+++ weather.py
@@ -4,7 +4,7 @@ from dotenv import load_dotenv
 from flask_migrate import Migrate, upgrade
 
 from app import create_app, db
-from app.models import Hour, Day
+from app.models import Day, Hour, Locality
 from app.weather import weather_factory
 from app.weather.base import WeatherAPI
 
@@ -18,17 +18,21 @@ migrate = Migrate(app, db)
 
 @app.shell_context_processor
 def make_shell_context():
-    return dict(db=db, Day=Day, Hour=Hour)
+    return dict(db=db, Day=Day, Hour=Hour, Locality=Locality)
 
 
 def fill_database(weater_api: 'WeatherAPI'):
-    from app.models import Day, Hour
-    days, hours = weater_api.collect_data()
-    Day.query.delete()
-    Hour.query.delete()
-    db.session.add_all(days)
-    db.session.add_all(hours)
-    db.session.commit()
+    from app.models import Day, Hour, Locality
+    forecast = weater_api.collect_data()
+    if forecast:
+        Day.query.delete()
+        Hour.query.delete()
+        Locality.query.delete()
+        for locality, f in forecast.items():
+            db.session.add(locality)
+            db.session.add_all(f['daily_forecast'])
+            db.session.add_all(f['hourly_forecast'])
+        db.session.commit()
 
 
 @app.cli.command()
@@ -37,7 +41,7 @@ def forecast():
     if api_name == 'TuTiempoAPI':
         api = weather_factory(api_name,
                               api_key=os.getenv('WEATHER_API_KEY'),
-                              geolocation=os.getenv('WEATHER_GEOLOCATION').split(','),
+                              locations=os.getenv('WEATHER_LOCATIONS').split(','),
                               language=os.getenv('WEATHER_LANGUAGE'))
     else:
         exit("Invalid Weather API")
